LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/visual-scrollIntoView-003-expected.txt


FAIL Element.scrollIntoView scrolls visually to an element in nested position: fixed elements assert_greater_than: expected a number greater than 1 but got 1
